==> analysis/2_scRNA_processing_part2.Rmd <==
---
title: "2_scRNA_processing_part2"
author: "Mette Ludwig"
date: "2021-02-05"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

```{r, warning=FALSE, message=FALSE}
library(WGCNA)
library(flashClust)
library(dplyr)
library(Matrix)
library(Seurat)
library(gProfileR)
library(scales)
library(ggplot2)
allowWGCNAThreads(20)
```

# Load AP Seurat object
```{r}
AP <- readRDS("/projects/mludwig/Ludwig-2021/output/Seurat_objects/AP_Seurat_obj_tmp1.rds")
```

# Perform WGCNA 
```{r}
DefaultAssay(AP) <- "RNA"
AP <- FindVariableFeatures(object = AP, selection.method = "vst", 
                           nfeatures = 5000, verbose = F)
AP <- ScaleData(object = AP, verbose = F) 
AP <- RunPCA(object = AP, npcs = 30, verbose = F)
```

```{r}
# Find top 5000 variable genes
pca.genes <- AP@reductions$pca@feature.loadings[,c(1:30)]
pca.genes <- abs(pca.genes)
pca.genes <- apply(pca.genes, 1, function(x) max(x))
pca.genes <- names(pca.genes[order(-pca.genes)])[1:5000]
ExprMat <- t(as.matrix(AP@assays$RNA[rownames(AP@assays$RNA) %in% pca.genes,]))
gene.names <- colnames(ExprMat)
```

```{r}
# Choose the soft threshold power beta
# Calculate potential thresholds
powers = c(c(1:10), seq(from = 12, to = 20, by = 2))

sft <- pickSoftThreshold(ExprMat, dataIsExpr = T, powerVector = powers, 
                         corFnc = "cor", blockSize = ncol(ExprMat),
                         corOptions = list(use = "p"), networkType = "signed")

# Plot scale independence
plot(sft$fitIndices[,1], -sign(sft$fitIndices[,3])*sft$fitIndices[,2], 
     xlab = "Soft Threshold (power)",
     ylab = "Scale Free Topology Model Fit, signed R^2", 
     type="n", main = paste("Scale independence"));
text(sft$fitIndices[,1], -sign(sft$fitIndices[,3])*sft$fitIndices[,2], 
     labels = powers, cex = 0.9, col = "red");
abline(h=0.80, col="red")
```


```{r}
softPower <- 7

# Compute TOM dissimilarity 
TOM <- TOMsimilarityFromExpr(ExprMat, networkType = "signed", TOMType = "signed", 
                             power = softPower);
colnames(TOM) <- rownames(TOM) <- gene.names
dissTOM <- 1-TOM

# Cluster the TOM dissimilarity matrix
geneTree <- flashClust(as.dist(dissTOM), method="complete") 
plot(geneTree, xlab="", sub="", cex=0.7, labels = FALSE)

# Tree cutting
minModuleSize <- 20
dynamicMods <- cutreeDynamic(dendro = geneTree, distM = dissTOM, method = "hybrid", 
                             deepSplit = 3, 
                             pamRespectsDendro = T, minClusterSize = minModuleSize)
dynamicColors <- labels2colors(dynamicMods) 

# Plot modules
plotDendroAndColors(geneTree, dynamicColors, "Dynamic Tree Cut", dendroLabels = FALSE, 
                    hang = 0.03, addGuide = TRUE, guideHang = 0.05, 
                    main = "Gene dendrogram and module colors") 

# Compute module Eigengenes
ExprMat <- as.matrix(ExprMat) 
MEs <- moduleEigengenes(ExprMat, dynamicColors)$eigengenes  

# Merge modules that have a Pearson correlation > 0.75
# Calculate dissimilarity of module eigengenes
MEDiss = 1-cor(MEs)
# Cluster module eigengenes
METree = flashClust(as.dist(MEDiss), method = "average");

# Plot the result
plot(METree, main = "Clustering of module eigengenes",
     xlab = "", sub = "")
MEDissThres = 0.75
abline(h=MEDissThres, col = "red")

# Merge modules
merge <- mergeCloseModules(ExprMat, dynamicColors, MEs=NULL, useSets=NULL, 
                           impute = T, checkDataFormat=T, 
                           corFnc = cor, corOptions=list(use='p'), useAbs = F, 
                           cutHeight = MEDissThres, verbose = 3)

# The merged module colors
mergedColors = merge$colors;

# Eigengenes of the new merged modules
mergedMEs = merge$newMEs

plotDendroAndColors(geneTree, cbind(dynamicColors, mergedColors),
                    c("Dynamic Tree Cut", "Merged dynamic"),
                    dendroLabels = FALSE, hang = 0.03,
                    addGuide = TRUE, guideHang = 0.05)

rownames(mergedMEs) <- rownames(ExprMat) 


# Build network object 
gene.assignments <- data.frame(mergedColors)
rownames(gene.assignments) <- colnames(ExprMat) # assign gene names as row names
module.names <- gsub("ME", "", colnames(mergedMEs)) 

Network = list() 
for (i in 1:length(module.names)){
  genes <- which(gene.assignments == module.names[i])
  Network[[i]] <- rownames(gene.assignments)[genes]
}
names(Network) <- module.names
```

```{r}
resultsGO = data.frame()

for (m in unique(mergedColors)) {

  ## GO Enrichment
  query = rownames(t(ExprMat))[mergedColors==m]
  go = gprofiler(query, max_set_size = 500, src_filter = "GO", 
                 correction_method = "fdr", min_isect_size = 3,
                 hier_filtering = "none")

  go = go[order(go$p.value),]
  num_to_plot = 10
  par(oma=c(0,10,0,0))
  bp = barplot(-log10(as.numeric(go$p.value[num_to_plot:1])),
               main=paste("\n\n\n\n\n",m, "- GO, KEGG"), horiz=T, yaxt='n',
               col=m, xlab='-log10(FDR)\n',cex.main=0.7, cex.axis = .7, xlim = c(0,80))
  axis(2, at=bp, labels=go$term.name[num_to_plot:1], tick=FALSE, las=2, cex.axis=.8);
  abline(v=-log10(0.01), col="red", lwd=2,lty=2)
  
  m2 <- as.data.frame(rep(m, nrow(go)))

  resultsGO = rbind(resultsGO, cbind(m2, go))
}
```

# Remove likely doublets 
```{r}
AP@meta.data[, paste0("M", seq(1:ncol(mergedMEs)))] <- mergedMEs

# Oligodendrocyte module
FeaturePlot(object = AP, features = "M5", cols= c("grey", "blue"), 
            reduction = "umap")

# Astrocyte module
FeaturePlot(object = AP, features = "M7", cols= c("grey", "blue"), 
            reduction = "umap")

# Microglia module
FeaturePlot(object = AP, features = "M10", cols= c("grey", "blue"), 
            reduction = "umap")
```

```{r, warning=FALSE}
# Define cutoffs 
M5.cutoff <- quantile(AP$M5[which(AP$cell.type == "Oligodendrocytes")], prob = 0.05)
M7.cutoff <- quantile(AP$M7[which(AP$cell.type == "Astrocytes")], prob = 0.05)
M10.cutoff <- quantile(AP$M10[which(AP$cell.type == "Microglia")], prob = 0.05)

# Remove cells that express modules specific to other cell types
AP$remove <- "No"
AP$remove[which(AP$M5 > M5.cutoff & !AP$cell.type %in% c("Oligodendrocytes", "OPCs"))] <- "Yes"
AP$remove[which(AP$M7 > M7.cutoff & !AP$cell.type %in% c("Astrocytes", 
                                                         "Tanycyte-like cells", 
                                                         "Ependymal cells", 
                                                         "Endothelial cells",
                                                         "VLMCs"))] <- "Yes"
AP$remove[which(AP$M10 > M10.cutoff & !AP$cell.type == "Microglia")] <- "Yes"

DimPlot(AP, reduction = "umap", label = T, group.by = "remove") + 
  theme(legend.position = "none") +
  xlab("UMAP 1") + ylab("UMAP 2")
table(AP$cell.type[which(AP$remove == "Yes")])
```

# Visualization
```{r}
AP <- subset(AP, subset = remove == "Yes", invert = T)

DefaultAssay(AP) <- "integrated"
AP <- ScaleData(object = AP, verbose = F) 
AP <- RunPCA(object = AP, npcs = 30, verbose = F)
AP <- RunUMAP(object = AP, reduction = "pca", 
    dims = 1:10)

DimPlot(AP, reduction = "umap", label = T, group.by = "cell.type") + 
  theme(legend.position = "none") +
  xlab("UMAP 1") + ylab("UMAP 2")
```

```{r}
saveRDS(AP, 
        file = "/projects/mludwig/Ludwig-2021/output/Seurat_objects/AP_Seurat_obj_tmp2.rds")
```
